services/sqs-observer/requirements.txt
kafka-python
boto3
